Completed ToDo's
Note: This code is very buggy and required a phone with a lot of memory to run - it will not run on Nexus 5 (Android 6.0).  On some phones the image must be taken in landscape mode.  Changing orientation after taking image can cause crash.

Added class Emojifier
Added method to detect number of faces in image (and log them)